Moved the easing functions into separated files for modularization
