+++ src/ease-in-cubic.js
@@ -1,0 +1,4 @@
+// Accelerating from zero velocity
+export default function easeInCubic( t ) {
+    return t * t * t;
+}
