+++ src/ease-in-elastic.js
@@ -1,0 +1,19 @@
+// Bounces slowly then quickly to finish
+export default function easeInElastic( t, magnitude = 0.7 ) {
+
+    if( t === 0 || t === 1 ) {
+        return t;
+    }
+
+    const scaledTime = t / 1;
+    const scaledTime1 = scaledTime - 1;
+
+    const p = 1 - magnitude;
+    const s = p / ( 2 * Math.PI ) * Math.asin( 1 );
+
+    return -(
+        Math.pow( 2, 10 * scaledTime1 ) *
+        Math.sin( ( scaledTime1 - s ) * ( 2 * Math.PI ) / p )
+    );
+
+}
