+++ src/ease-in-expo.js
@@ -1,0 +1,10 @@
+// Accelerate exponentially until finish
+export default function easeInExpo( t ) {
+
+    if( t === 0 ) {
+        return 0;
+    }
+
+    return Math.pow( 2, 10 * ( t - 1 ) );
+
+}
