+++ src/ease-in-out-circ.js
@@ -1,0 +1,13 @@
+// Fast increase in velocity, fast decrease in velocity
+export default function easeInOutCirc( t ) {
+
+    const scaledTime = t * 2;
+    const scaledTime1 = scaledTime - 2;
+
+    if( scaledTime < 1 ) {
+        return -0.5 * ( Math.sqrt( 1 - scaledTime * scaledTime ) - 1 );
+    }
+
+    return 0.5 * ( Math.sqrt( 1 - scaledTime1 * scaledTime1 ) + 1 );
+
+}
