+++ src/ease-in-out-cubic.js
@@ -1,0 +1,4 @@
+// Acceleration until halfway, then deceleration
+export default function easeInOutCubic( t ) {
+    return t < 0.5 ? 4 * t * t * t : ( t - 1 ) * ( 2 * t - 2 ) * ( 2 * t - 2 ) + 1;
+}
