+++ src/ease-in-out-quad.js
@@ -1,0 +1,4 @@
+// Acceleration until halfway, then deceleration
+export default function easeInOutQuad( t ) {
+    return t < 0.5 ? 2 * t * t : - 1 + ( 4 - 2 * t ) * t;
+}
