+++ src/ease-in-out-quart.js
@@ -1,0 +1,5 @@
+// Acceleration until halfway, then deceleration
+export default function easeInOutQuart( t ) {
+    const t1 = t - 1;
+    return t < 0.5 ? 8 * t * t * t * t : 1 - 8 * t1 * t1 * t1 * t1;
+}
