+++ src/ease-in-out-quint.js
@@ -1,0 +1,5 @@
+// Acceleration until halfway, then deceleration
+export default function easeInOutQuint( t ) {
+    const t1 = t - 1;
+    return t < 0.5 ? 16 * t * t * t * t * t : 1 + 16 * t1 * t1 * t1 * t1 * t1;
+}
