+++ src/ease-in-out-sine.js
@@ -1,0 +1,4 @@
+// Slight acceleration at beginning and slight deceleration at end
+export default function easeInOutSine( t ) {
+    return -0.5 * ( Math.cos( Math.PI * t ) - 1 );
+}
