+++ src/ease-in-quad.js
@@ -1,0 +1,4 @@
+// Accelerating from zero velocity
+export default function easeInQuad( t ) {
+    return t * t;
+}
