+++ src/ease-in-quart.js
@@ -1,0 +1,4 @@
+// Accelerating from zero velocity
+export default function easeInQuart( t ) {
+    return t * t * t * t;
+}
