+++ src/ease-in-quint.js
@@ -1,0 +1,4 @@
+// Accelerating from zero velocity
+export default function easeInQuint( t ) {
+    return t * t * t * t * t;
+}
