+++ src/ease-in-sine.js
@@ -1,0 +1,4 @@
+// Slight acceleration from zero to full speed
+export default function easeInSine( t ) {
+    return -1 * Math.cos( t * ( Math.PI / 2 ) ) + 1;
+}
