+++ src/ease-out-bounce.js
@@ -1,0 +1,27 @@
+// Bounce to completion
+export default function easeOutBounce( t ) {
+
+    const scaledTime = t / 1;
+
+    if( scaledTime < ( 1 / 2.75 ) ) {
+
+        return 7.5625 * scaledTime * scaledTime;
+
+    } else if( scaledTime < ( 2 / 2.75 ) ) {
+
+        const scaledTime2 = scaledTime - ( 1.5 / 2.75 );
+        return ( 7.5625 * scaledTime2 * scaledTime2 ) + 0.75;
+
+    } else if( scaledTime < ( 2.5 / 2.75 ) ) {
+
+        const scaledTime2 = scaledTime - ( 2.25 / 2.75 );
+        return ( 7.5625 * scaledTime2 * scaledTime2 ) + 0.9375;
+
+    } else {
+
+        const scaledTime2 = scaledTime - ( 2.625 / 2.75 );
+        return ( 7.5625 * scaledTime2 * scaledTime2 ) + 0.984375;
+
+    }
+
+}
