+++ src/ease-out-expo.js
@@ -1,0 +1,10 @@
+// Initial exponential acceleration slowing to stop
+export default function easeOutExpo( t ) {
+
+    if( t === 1 ) {
+        return 1;
+    }
+
+    return ( -Math.pow( 2, -10 * t ) + 1 );
+
+}
