+++ src/ease-out-quad.js
@@ -1,0 +1,4 @@
+// Decelerating to zero velocity
+export default function easeOutQuad( t ) {
+    return t * ( 2 - t );
+}
