+++ src/ease-out-quart.js
@@ -1,0 +1,5 @@
+// Decelerating to zero velocity
+export default function easeOutQuart( t ) {
+    const t1 = t - 1;
+    return 1 - t1 * t1 * t1 * t1;
+}
