+++ src/ease-out-sine.js
@@ -1,0 +1,4 @@
+// Slight deceleration at the end
+export default function easeOutSine( t ) {
+    return Math.sin( t * ( Math.PI / 2 ) );
+}
