+++ src/linear.js
@@ -1,0 +1,4 @@
+// No easing, no acceleration
+export default function linear( t ) {
+    return t;
+}
